사용 함수 ui로 이전
