+++ pages/ctf01.py
@@ -3,9 +3,6 @@
 import streamlit as st
 from utils.llm_utils import ctf01_llm_ask, ctf01_llm_flag
 from utils.ui import render_main_header, render_flag_sub
-# from utils.auth import get_user, require_login
-
-# user = require_login()
 
 render_main_header()
