+++ utils/ui.py
@@ -1,7 +1,8 @@
 # utils/ui.py
 import streamlit as st
 import csv
-
+import io
+import chardet
 
 # 메인으로 돌아가는 버튼
 def render_main_header():
@@ -35,18 +36,7 @@ def render_flag_sub(flag, challenge_id: str):
         else:
             st.error("❌ 틀렸습니다.")
 
-
-# 업로드된 .txt파일에서 텍스트 추출 함수
-def extract_text(uploaded_file):
-    """업로드된 .txt파일에서 텍스트 추출 함수"""
-    try:
-        text = uploaded_file.read().decode("utf-8")
-        return text.strip()
-    except Exception as e:
-        return f"❌ 파일 처리 중 오류 발생: {e}"
-
-
-# ctf01 사용하는 scv파일 읽기 함수
+# CTF01 - csv파일 읽기 함수
 def csv_read_func(file_path):
     with open(file_path, "r", encoding="utf-8") as f:
         reader = csv.DictReader(f)
@@ -56,3 +46,46 @@ def csv_read_func(file_path):
                 f"- 주문자: {row['주문자']} / 주문코드: {row['주문코드']} / 상품명: {row['상품명']} / 가격: {int(row['가격']):,}원"
             )
         return "\n".join(lines)
+
+# CTF04 - CSV 정보 읽는 함수
+def generate_prompt_from_csv(csv_text):
+    f = io.StringIO(csv_text)
+    reader = csv.reader(f)
+    lines = []
+    for row in reader:
+        lines.extend(row)
+    cleaned = [line.strip() for line in lines if line.strip()]
+    return "\n".join(cleaned)
+
+# CTF04 - raw_data를 다양한 인코딩으로 디코딩 시도
+def try_decode(raw_data) -> tuple[str | None, str]:
+    """raw_data를 다양한 인코딩으로 디코딩 시도"""
+    encodings_to_try = ["utf-8", "cp949", "euc-kr", "iso-8859-1"]
+
+    for enc in encodings_to_try:
+        try:
+            text = raw_data.decode(enc)
+            return text, f"{enc}"
+        except UnicodeDecodeError:
+            continue
+
+    # chardet 자동 인코딩 감지
+    result = chardet.detect(raw_data)
+    encoding = result.get("encoding")
+    confidence = result.get("confidence", 0)
+
+    if encoding:
+        try:
+            text = raw_data.decode(encoding)
+            return text, f"{encoding} (자동감지, 신뢰도 {confidence*100:.1f}%)"
+        except:
+            pass
+
+# CTF08 - 업로드된 .txt파일에서 텍스트 추출 함수
+def extract_text(uploaded_file):
+    """업로드된 .txt파일에서 텍스트 추출 함수"""
+    try:
+        text = uploaded_file.read().decode("utf-8")
+        return text.strip()
+    except Exception as e:
+        return f"❌ 파일 처리 중 오류 발생: {e}"
